requirements.txt
django~=4.0.6
djangorestframework~=3.13.1
pandas~=1.4.3
numpy~=1.23.1
requests~=2.26
pytest~=7.1.3
pytest-django~=4.5.2
awsebcli~=3.20.3
pulp~=2.6.0